week-2/crypto-etl-pipeline_refinements/requirements.txt
pandas==3.0.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.5